move getAdminLevelHierarchy func definition to v2-events/utils
