+++ packages/client/src/v2-events/features/events/registered-fields/Address.tsx
@@ -37,6 +37,7 @@ import { getOfflineData } from '@client/offline/selectors'
 import { useLocations } from '@client/v2-events/hooks/useLocations'
 import { IAdminStructureItem } from '@client/utils/referenceApi'
 import { getUserDetails } from '@client/profile/profileSelectors'
+import { getAdminLevelHierarchy } from '@client/v2-events/utils'
 
 // ADDRESS field may not contain another ADDRESS field
 type FieldConfigWithoutAddress = Exclude<
@@ -128,51 +129,6 @@ function isDomesticAddress() {
     not(createFieldCondition('country').isUndefined()),
     createFieldCondition('addressType').isEqualTo(AddressType.DOMESTIC)
   )
-}
-
-type OutputMode = 'withIds' | 'withNames'
-/*
-Function to traverse the administrative leverl hierarchy from an arbitrary / leaf point
-*/
-export function getAdminLevelHierarchy(
-  locationUuid: string | undefined,
-  locations: Location[],
-  adminStructure: string[],
-  outputMode: OutputMode = 'withIds'
-) {
-  // Collect location objects from leaf to root
-  const collectedLocations: Location[] = []
-
-  let current = locationUuid
-    ? locations.find((l) => l.id === locationUuid.toString())
-    : null
-
-  while (current) {
-    collectedLocations.push(current)
-    if (!current.parentId) {
-      break
-    }
-    const parentId = current.parentId
-    current = locations.find((l) => l.id === parentId)
-  }
-
-  // Reverse so root is first, leaf is last
-  collectedLocations.reverse()
-
-  // Map collected locations to the provided admin structure
-  const hierarchy: Partial<Record<string, string>> = {}
-  for (
-    let i = 0;
-    i < adminStructure.length && i < collectedLocations.length;
-    i++
-  ) {
-    hierarchy[adminStructure[i]] =
-      outputMode === 'withNames'
-        ? collectedLocations[i].name
-        : collectedLocations[i].id
-  }
-
-  return hierarchy
 }
 
 function generateAdminStructureFields(
